fix: refactored frankenstein modules into daemons
