+++ projects/angular-simple-oidc/src/lib/session-management/session-management.module.ts
@@ -1,12 +1,5 @@
-import { NgModule, OnDestroy, ModuleWithProviders } from '@angular/core';
-import { Subject, of } from 'rxjs';
-import { switchMap, takeUntil, tap, catchError, take, map } from 'rxjs/operators';
-import { TokensReadyEvent } from '../auth.events';
+import { NgModule,  ModuleWithProviders } from '@angular/core';
 import { SessionCheckService } from './session-check.service';
-import { SessionChangedEvent, SessionTerminatedEvent } from './events';
-import { TokenStorageService } from '../token-storage.service';
-import { SimpleOidcError } from 'angular-simple-oidc/core';
-import { filterInstanceOf } from 'angular-simple-oidc/operators';
 import { AuthorizeEndpointSilentClientService } from './authorize-endpoint-silent-client.service';
 import {
   SESSION_MANAGEMENT_CONFIG,
@@ -14,87 +7,36 @@ import {
   SESSION_MANAGEMENT_CONFIG_SERVICE_PROVIDER
 } from './providers';
 import { SessionManagementConfig } from './models';
-import { EventsService, SimpleOidcInfoEvent } from 'angular-simple-oidc/events';
+import { SessionCheckDaemonService } from './session-check-daemon.service';
+import { SilentFlowDaemonService } from './silent-flow-daemon.service';
 
 /**
  * Implements Session Checks according to Session Management
  * https://openid.net/specs/openid-connect-session-1_0.html
  */
-// TODO: Add Angular decorator.
 @NgModule({
-  imports: [],
   providers: [
     SessionCheckService,
     AuthorizeEndpointSilentClientService,
+    SessionCheckDaemonService,
+    SilentFlowDaemonService,
   ],
-  declarations: [],
 })
-export class SessionManagementModule implements OnDestroy {
-
-  protected readonly destroyedSubject = new Subject();
-
+export class SessionManagementModule {
   constructor(
-    protected readonly sessionCheck: SessionCheckService,
-    protected readonly events: EventsService,
-    protected readonly authorizeSilentClient: AuthorizeEndpointSilentClientService,
-    protected readonly tokenStorage: TokenStorageService,
+    protected readonly sessionCheckDaemonService: SessionCheckDaemonService,
+    protected readonly silentFlowDaemonService: SilentFlowDaemonService,
   ) {
-
-    const sessionChanged$ = this.events.events$
-      .pipe(filterInstanceOf(SessionChangedEvent));
-
-    // Start session checks when we receive tokens.
-    this.events.events$
-      .pipe(
-        filterInstanceOf(TokensReadyEvent),
-        switchMap(() =>
-          this.sessionCheck.startSessionCheck()
-            .pipe(takeUntil(sessionChanged$))
-        ),
-        takeUntil(this.destroyedSubject),
-      )
-      .subscribe();
-
-    // When the RP detects a session state change,
-    // it SHOULD first try a prompt=none request within an iframe to obtain
-    // a new ID Token and session state, sending the old ID Token as the id_token_hint.
-    // If the RP receives an ID token for the same End-User, it SHOULD simply update
-    // the value of the session state. If it doesn't receive an ID token or receives
-    // an ID token for another End-User, then it needs to handle this case as a
-    // logout for the original End-User.
-    const silentCodeFlow$ = this.tokenStorage.currentState$
-      .pipe(
-        take(1),
-        switchMap((previousState) => this.authorizeSilentClient.startCodeFlowInIframe()
-          .pipe(map(result => ({ result, previousState })))
-        ),
-        tap(({ result, previousState }) => {
-          const newToken = result.decodedIdToken;
-          const previousToken = previousState.decodedIdentityToken;
-          if (previousToken && newToken.sub !== previousToken.sub) {
-            this.events.dispatch(new SessionTerminatedEvent({ previousToken, newToken }));
-          }
-        }),
-        catchError((error: SimpleOidcError) => {
-          this.events.dispatch(new SessionTerminatedEvent({ error: error.context }));
-          return of(null);
-        })
-      );
-
-    this.events.events$
-      .pipe(
-        filterInstanceOf(SessionChangedEvent),
-        tap(() => this.events.dispatch(new SimpleOidcInfoEvent('Session has changed. Starting silent code flow.'))),
-        switchMap(() => silentCodeFlow$),
-        takeUntil(this.destroyedSubject)
-      )
-      .subscribe();
+    this.sessionCheckDaemonService.startDaemon();
+    this.silentFlowDaemonService.startDaemon();
   }
 
   /**
    * Should be called once on your Angular Root Application Module
    */
   public static forRoot(config?: SessionManagementConfig): ModuleWithProviders<SessionManagementModule> {
+    // TODO: add sanity checks with forRoot + @Optional()
+
     return {
       ngModule: SessionManagementModule,
       providers: [
@@ -106,10 +48,5 @@ export class SessionManagementModule implements OnDestroy {
         SESSION_MANAGEMENT_CONFIG_INITIALIZER,
       ]
     };
-  }
-
-  public ngOnDestroy() {
-    this.destroyedSubject.next();
-    this.destroyedSubject.complete();
   }
 }
